# Conflicts:
#	.gitignore
#	example/pubspec.lock
#	example/windows/AppConfiguration.props
#	example/windows/Runner.sln
#	example/windows/Runner.vcxproj
#	example/windows/flutter/.template_version
#	example/windows/scripts/bundle_assets_and_deps.bat
#	lib/Command.dart
#	lib/Response.dart
#	lib/flutter_usb.dart
#	pubspec.lock
#	pubspec.yaml
#	windows/PluginInfo.props
#	windows/flutter_usb_plugin.cpp
#	windows/plugin.vcxproj
#	windows/plugin.vcxproj.filters
#	windows/response.h

+++ lib/Response.dart
@@ -1,6 +1,8 @@
 class Response {
+import 'dart:typed_data';
+
   String result;
-  int outDataLength;
+double outDataLength;
   List<int> inData;
 
   Response(this.result, this.outDataLength, this.inData);
